+++ apps/frontend/src/app/[locale]/lems/(volunteer)/audience-display/components/scoreboard/table/scores-table-header.tsx
